requirements.txt
streamlit
tensorflow-cpu
pillow
pandas
matplotlib
gtts
numpy